arkla-backend/requirements.txt
# ARKLA Backend Dependencies
# Generated from MCP v1.0

fastapi>=0.104.0
uvicorn>=0.24.0
google-genai>=0.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
pandas>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
regex>=2023.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
openpyxl>=3.1.0
